code/requirements.txt
absl-py==1.0.0
adabound==0.0.5
albumentations==1.1.0
argon2-cffi @ file:///tmp/build/80754af9/argon2-cffi_1613037097816/work
asttokens==2.0.5
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1620387926260/work
backcall==0.2.0
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
batchgenerators==0.23
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1629908509068/work
blis==0.7.5
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.10.8
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1636046063618/work
charset-normalizer==2.0.9
click==8.0.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
cycler==0.11.0
cymem==2.0.6
dalib==0.2
dataclasses==0.6
debugpy==1.5.1
decorator==5.1.1
defusedxml @ file:///home/conda/feedstock_root/build_artifacts/defusedxml_1615232257335/work
dominate==2.6.0
easydict==1.9
efficientnet-pytorch==0.7.1
einops==0.4.1
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
executing==0.8.3
fastai==2.5.3
fastcore==1.3.27
fastdownload==0.0.5
fastprogress==1.0.0
fonttools==4.28.3
future==0.18.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
h5py==3.6.0
idna==3.3
imageio==2.13.2
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1636431522037/work
importlib-resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1635615662634/work
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets @ file:///home/conda/feedstock_root/build_artifacts/ipywidgets_1631590360471/work
jedi==0.18.1
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1636510082894/work
joblib==1.1.0
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema-meta_1639599819103/work
jupyter @ file:///home/conda/feedstock_root/build_artifacts/jupyter_1637233111631/work
jupyter-client==7.1.2
jupyter-console @ file:///home/conda/feedstock_root/build_artifacts/jupyter_console_1616560109969/work
jupyter-core==4.9.2
jupyterlab-widgets @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_widgets_1631590465624/work
kiwisolver==1.3.2
langcodes==3.3.0
lxml==4.8.0
Markdown==3.3.6
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1621455677251/work
matplotlib==3.5.0
matplotlib-inline==0.1.3
MedPy==0.4.0
mistune @ file:///home/conda/feedstock_root/build_artifacts/mistune_1624941302223/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
murmurhash==1.0.6
nbconvert @ file:///home/conda/feedstock_root/build_artifacts/nbconvert_1631125953237/work
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1617383142101/work
nest-asyncio==1.5.4
networkx==2.6.3
nibabel==3.2.1
notebook @ file:///home/conda/feedstock_root/build_artifacts/notebook_1637122013008/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634095647912/work
oauthlib==3.1.1
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opencv-python==4.5.5.64
opencv-python-headless==4.5.4.60
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.3.5
pandocfilters @ file:///home/conda/feedstock_root/build_artifacts/pandocfilters_1631603243851/work
parso==0.8.3
pathy==0.6.1
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare==0.7.5
Pillow==8.4.0
preshed==3.0.6
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1635538335951/work
prompt-toolkit==3.0.28
protobuf==3.19.1
psutil==5.9.0
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydantic==1.8.2
pydicom @ file:///home/conda/feedstock_root/build_artifacts/pydicom_1633185052031/work
Pygments==2.11.2
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1636757021002/work
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1636110947380/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-docx==0.8.11
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
qpsolvers==1.7.2
qtconsole @ file:///home/conda/feedstock_root/build_artifacts/qtconsole_1639432716385/work
QtPy @ file:///home/conda/feedstock_root/build_artifacts/qtpy_1638557980349/work
quadprog==0.1.11
qudida==0.0.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.19.0
scikit-learn==1.0.1
scipy==1.7.3
Send2Trash @ file:///home/conda/feedstock_root/build_artifacts/send2trash_1628511208346/work
setuptools-scm==6.3.2
SimpleITK==2.1.1
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1623709665295/work
smart-open==5.2.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
stack-data==0.2.0
tb-nightly==2.8.0a20211224
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4.1
terminado @ file:///home/conda/feedstock_root/build_artifacts/terminado_1636052221351/work
testpath @ file:///home/conda/feedstock_root/build_artifacts/testpath_1621261527237/work
thinc==8.0.13
thop==0.0.31.post2005241907
threadpoolctl==3.0.0
tifffile==2021.11.2
timm==0.4.12
tomli==1.2.2
torch==1.10.1
torch-tb-profiler==0.3.1
torchaudio==0.7.0
torchsample==0.1.3
torchsummary==1.5.1
torchvision==0.11.2
tornado==6.1
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1632160078689/work
traitlets==5.1.1
typer==0.4.0
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.26.7
warp==1.0.4
wasabi==0.9.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension @ file:///home/conda/feedstock_root/build_artifacts/widgetsnbextension_1637174139311/work
yacs==0.1.8
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1633302054558/work
